Add batch_size_VS_resize() and add call as `__main__` in batchVSresize.py
Re-organize batchVSresize.py by adding a wrapper function batch_size_VS_resize that takes the configuration as argument and executes a single experiment. The idea is to use this as a building block and add wrappers/decorators on top of it in the future to integrate W&B and sweeps.