Add `Heatmap2d` class block
Implement `Heatmap2d` block to allow porting Keras models in a fastai-compatible way. In fact, fastai expects K channels as model output, where K is the number of object's classes (2 in our case, 0=background and 1=cell).
The intent in doing this is to make use of default fastai implementations for loss functions, learner utils and so on...
The attempted implementation replicates the previous Heatmap block and just concatenates 1-heatmap as classification scores for class 0.

#TODO: adapt ResUnet implementation