Add UpResidualBlock and comment out UpResNetBlock
Refactor UpResNetBlock as UpResidualBlock for better interpretability. In particular, the block now has a conv_path and an id_path, which are then added at the end. This is exactly the same structure of the ResidualBlock, so the implementation is now more coherent, the difference is that here the id_path is given by a ConvT2d + concat operations